load building location in rectangle boundary from OSM and display
When clicking on the Validate boundaries button, the max and min latitudes and longitudes
of the boundary are sent to Fastapi. A request is then made to overpass to get the json of
the area of interest. A Geojson is then created from the json and is returned to js. The Geojson
is then drawn on the map.